refactor: Rename `…lugins/aws/package/compile/events/s3/configSchema.js`
