+++ lib/plugins/aws/package/compile/events/s3/config-schema.js
