+++ lib/plugins/aws/provider.js
@@ -1144,7 +1144,7 @@ class AwsProvider {
             runtime: { $ref: '#/definitions/awsLambdaRuntime' },
             s3: {
               type: 'object',
-              additionalProperties: require('./package/compile/events/s3/configSchema'),
+              additionalProperties: require('./package/compile/events/s3/config-schema'),
             },
             stage: { $ref: '#/definitions/stage' },
             stackName: {
